requirements.txt
RPi.GPIO
discord
python-nmap
pygame
